+++ more/webassets/directives.py
@@ -6,10 +6,11 @@ from more.webassets.core import WebassetsApp
 from webassets import Bundle, Environment
 
 
-MISSING = object()
-
-
 class Asset(object):
+    """ Represents a registered asset which points to one or more files or
+    child-assets.
+
+    """
 
     __slots__ = ('name', 'assets', 'filters')
 
@@ -24,21 +25,14 @@ class Asset(object):
             and self.filters == self.filters
 
     @property
-    def is_single_file(self):
-        return len(self.assets) == 1 and '.' in self.assets[0]
-
-    @property
-    def path(self):
-        assert self.is_single_file
-        return self.assets[0]
-
-    @property
-    def extension(self):
-        if self.is_pure:
-            return self.assets[0].split('.')[-1]
-
-    @property
     def is_pure(self):
+        """ Returns True if this asset is "pure".
+
+        Pure assets are assets which consist of a single file or a set of
+        files which share one common extension.
+
+        """
+
         if self.is_single_file:
             return True
 
@@ -47,18 +41,50 @@ class Asset(object):
 
         return len(extensions) == 1 and None not in extensions
 
+    @property
+    def is_single_file(self):
+        """ Returns True if this repesents a single file asset. """
+        return len(self.assets) == 1 and '.' in self.assets[0]
+
+    @property
+    def path(self):
+        """ Returns the path to the single file asset if possible. """
+        assert self.is_single_file
+        return self.assets[0]
+
+    @property
+    def extension(self):
+        """ Returns the extension of this asset if it's a pure asset. """
+        if self.is_pure:
+            return self.assets[0].split('.')[-1]
+
 
 class WebassetRegistry(object):
+    """ A registry managing webasset bundles registered through directives. """
 
     def __init__(self):
+
+        #: A list of all paths which should be searched for files (in order)
         self.paths = []
+
+        #: The default filters for extensions. Each extension has a webassets
+        #: filter string associated with it. (e.g. {'js': 'rjsmin'})
         self.filters = {}
-        self.files = {}
+
+        #: :class:`Asset` objects keyed by their name
         self.assets = {}
+
+        #: The output path for all bundles
         self.output_path = None
+
+        #: A cache of created bundles
         self.cached_bundles = {}
+
+        #: The url passed to the webasset environment
         self.url = 'assets'
-        self.environment_config = {}
+
+        #: more.webasset only publishes js/css files - other file extensions
+        #: need to be compiled into either and mapped accordingly
         self.mapping = {
             'coffee': 'js',
             'dust': 'js',
@@ -71,13 +97,26 @@ class WebassetRegistry(object):
         }
 
     def register_path(self, path):
+        """ Registers the given path as a path to be searched for files.
+
+        The paths are prepended, so each new path has higher precedence than
+        all the already registered paths.
+
+        """
         assert os.path.isabs(path), "absolute paths only"
         self.paths.insert(0, os.path.normpath(path))
 
     def register_filter(self, name, filter):
+        """ Registers a filter, overriding any existing filter of the same
+        name.
+
+        """
         self.filters[name] = filter
 
     def register_asset(self, name, assets, filters=None):
+        """ Registers a new asset.
+
+        """
 
         assert '.' not in name, "asset names may not contain dots ({})".format(
             name
@@ -107,6 +146,8 @@ class WebassetRegistry(object):
                 assert asset in self.assets, "unknown asset {}".format(asset)
 
     def find_file(self, name):
+        """ Searches for the given file by name using the current paths. """
+
         if os.path.isabs(name):
             return name
 
@@ -126,6 +167,11 @@ class WebassetRegistry(object):
         raise LookupError("Could not find {} in paths".format(name))
 
     def merge_filters(self, *filters):
+        """ Takes a list of filters and merges them.
+
+        The last filter has the highest precedence.
+
+        """
         result = {}
 
         for filter in filters:
@@ -135,6 +181,8 @@ class WebassetRegistry(object):
         return result
 
     def get_bundles(self, name, filters=None):
+        """ Yields all the bundles for the given name (an asset). """
+
         assert name in self.assets, "unknown asset {}".format(name)
         assert self.output_path, "no webasset_output path set"
 
@@ -163,6 +211,8 @@ class WebassetRegistry(object):
                     yield bundle
 
     def get_environment(self):
+        """ Returns the webassets environment, registering all the bundles. """
+
         env = Environment(
             directory=self.output_path,
             load_path=self.paths,
@@ -203,6 +253,26 @@ class WebassetRegistry(object):
 
 @WebassetsApp.directive('webasset_path')
 class WebassetPath(Action):
+    """ Registers a path with more.webassets.
+
+    Registered paths are searched for assets registered::
+
+        @App.webasset_path()
+        def get_asset_path():
+            return 'assets/js'  # relative to the directory of the code file
+
+        @App.webasset('jquery.js')
+        def get_jquery_asset():
+            yield 'jquery.js'  # expected to be at assets/js/jquery.js
+
+    Registered paths can be accumulated, that is you can't override existing
+    paths, you can just add new paths which take precedence
+    (think ``PATH=/new/path:$PATH``).
+
+    Therefore paths registered first are searched last and paths registered
+    by a parent class are search after paths registered by the child class.
+
+    """
 
     config = {
         'webasset_registry': WebassetRegistry
@@ -218,11 +288,30 @@ class WebassetPath(Action):
             return os.path.join(os.path.dirname(self.code_info.path), path)
 
     def perform(self, obj, webasset_registry):
+        path = self.absolute_path(obj())
+        assert os.path.isdir(path), "'{}' does not exist".format(path)
+
         webasset_registry.register_path(self.absolute_path(obj()))
 
 
 @WebassetsApp.directive('webasset_filter')
 class WebassetFilter(Action):
+    """ Registers a default filter for an extension.
+
+    Filters are strings interpreted by `webasset`::
+
+        @App.webasset_filter('js')
+        def get_js_filter():
+            return 'rjsmin'
+
+        @App.webasset_filter('scss')
+        def get_scss_filter():
+            return 'pyscss'
+
+    For a list of available filters see
+    `<http://webassets.readthedocs.org/en/latest/builtin_filters.html>`_.
+
+    """
 
     group_class = WebassetPath
 
@@ -238,6 +327,15 @@ class WebassetFilter(Action):
 
 @WebassetsApp.directive('webasset_output')
 class WebassetOutput(Action):
+    """ Sets the output path for all bundles.
+
+    For example::
+
+        @App.webasset_output()
+        def get_output_path():
+            return 'assets/bundles'
+
+    """
 
     group_class = WebassetPath
 
@@ -250,6 +348,23 @@ class WebassetOutput(Action):
 
 @WebassetsApp.directive('webasset_mapping')
 class WebassetMapping(Action):
+    """ Maps an extension to either css or js.
+
+    You usually don't have to use this, as more.webassets comes with default
+    values. If you do, please open an issue so your mapping may be added
+    to more.webassets.
+
+    Example::
+
+        @App.webasset_mapping('jsx')
+        def get_jsx_mapping():
+            return 'js'
+
+        @App.webasset_mapping('less')
+        def get_jsx_mapping():
+            return 'css'
+
+    """
 
     group_class = WebassetPath
 
@@ -265,6 +380,17 @@ class WebassetMapping(Action):
 
 @WebassetsApp.directive('webasset_url')
 class WebassetUrl(Action):
+    """ Defines the url under which the bundles should be served.
+
+    Passed to the webasset environment, this is basically a url path prefix::
+
+        @App.webasset_url()
+        def get_webasset_url():
+            return 'my-assets'
+
+    Defaults to 'assets'.
+
+    """
 
     group_class = WebassetPath
 
@@ -277,6 +403,53 @@ class WebassetUrl(Action):
 
 @WebassetsApp.directive('webasset')
 class Webasset(Action):
+    """ Registers an asset which may then be included in the page.
+
+    For example::
+
+        @App.webasset('tooltip')
+        def get_tooltip_asset():
+            yield 'tooltip.js'
+            yield 'tooltip.css'
+
+    Assets may be included by using
+    :meth:`more.webassets.core.IncludeRequest.include`::
+
+        @App.view(model=Model)
+        def view_model(self, request):
+            request.include('tooltip')
+
+    Asset functions must be generators. They may include a mixed set of
+    assets. So you can freely mix css, js, less and so on. When you include
+    the asset by name those files are automatically put into appropriate
+    webasset bundles.
+
+    You may also define a custom filter which only applies to the registered
+    assets (they override the default filters registerd by `webasset_filter`)::
+
+        @App.webasset('tooltip', filters={'css': 'cssmin'})
+        def get_tooltip_asset():
+            yield 'tooltip.js'
+            yield 'tooltip.css'
+
+    Assets defined in a parent class may be overridden by using the same
+    name, or they may be reused. This means you can have applications which
+    define assets for you that you can reuse::
+
+        @BaseApp.webasset('react')
+        def get_react_asset():
+            yield 'react.js'
+
+        @App.webasset('widget')
+        def get_widget_asset():
+            yield 'react'
+            yield 'widget.jsx'
+
+    Note that webassets may not contain path separators. You're supposed to
+    register all paths which should be searched, and then you only work
+    with filenames.
+
+    """
 
     depends = [
         WebassetFilter,